Extract writeToFileUndoable
Create a function that can be reused to create a file without java.io
